session4/requirements.txt
openai>=1.40.0
chromadb>=0.5.0
tiktoken>=0.7.0
numpy>=1.26.0
python-dotenv>=1.0.0
